fix: improved parse() method using onvif namespaces (#12)
More cameras will be detected with this improvement.

BREAKING CHANGES: property "types: []" is no longer included in the device interface. Some devices were returning several responses and doesn't make sense to emit said responses for each device type. The point of the library is mostly to show cameras - this can be revisited later.